Add a different color for Typescript unit tests
I propose we visually differentiate Typescript unit test files `.spec.ts` with the same icon, but a different color.  Yellow was chosen as it's a part of the traffic light analogy that unit tests follow, but doesn't imply that the unit tests within the file are passing (green) or failing (red).