Changed Rust Icon
The rust icon seemed kind of noisy, so I took an SVG of Ferris the Rustacean and modified it for better appearance at a small size. I tried to make sure it fits within the style of the package as a whole while being distinctly 'Rust.'

Let me know if I need to do anything more to integrate it properly. I figured modifying the existing SVG file and changing the color in the mapping.less would be sufficient.